Sign in fixes
Needed a couple of more checks to handle multiple sign ins from the same person. Also the logout function is unique enough for open sessions that I moved it there.
