+++ sciencelabs/templates/session/student_attendance.html
@@ -30,7 +30,7 @@
             <br>
         {% endif %}
         <div class="centered margin-bottom-10">
-            <button type="submit" class="big-button btn btn-primary sign-in">Scan Card or Click Here to Sign In</button>
+            <button type="submit" class="big-button btn btn-primary sign-in">Scan ID or Click Here to Sign In</button>
         </div>
     </form>
 
@@ -82,7 +82,7 @@
                     };
                     $.post(url, scannedInput, function(success){
                         if(success != 'failed'){
-                           window.location.replace("{{ lab_url }}/session/checkin/{{ session_info.id }}/{{ session_info.hash }}/" + success);
+                            window.location.replace("{{ lab_url }}/session/checkin/{{ session_info.id }}/{{ session_info.hash }}/" + success);
                         }else{
                             input = "";
                         }
